treewide: clean up coro dependencies
