--- coro/inlude/gap/coro/recursive_generator.hpp
@@ -1,6 +1,0 @@
-// Copyright (c) 2024, Trail of Bits, Inc.
-
-#pragma once
-
-// To be backward compatible with the old include path
-#include "gap/core/recursive_generator.hpp"
